Inline small DuckLake writes with compatible storage (#272)
* feat: inline small DuckLake writes

Enable automatic small-write routing and complete scalar, nested, native-type, and index-aware inline storage. Integrate staged inline writes, strict maintenance reads, and flush behavior without changing the preceding transaction model.

* Fix inline storage compatibility and mutation fences

Preserve DuckLake temporal and nested binary encodings, and keep physical indexes opt-in without rewriting existing columns.

Fence direct and staged inline writes against partition changes, check current inline rows before UPDATE, reject incomplete change feeds, and count live rows during truncation. Cover mixed SQLite temporal storage and enable backend writer tests in CI.

* Fix writer CI regressions and reject non-empty CTAS

Align inline encoding assertions with text storage, complete the encrypted catalog fixture, and make the MySQL inline registry setup idempotent. Decode the PostgreSQL conflict sentinel using its actual INT4 type.

Reject materialized CTAS inputs before publishing metadata instead of discarding SELECT rows. Document CREATE TABLE followed by INSERT INTO ... SELECT as the supported path, and verify both rejection and empty-table creation.

* Keep data inlining opt-in

Default the catalog row limit to zero while inline UPDATE, rowid, CDC, SQL flush, and maintenance remain incomplete. Preserve explicit positive settings and keep inline regressions opted in.

Share the complete reserved inline-column check across writers, including the internal row and snapshot identifiers. Verify Parquet fallback, read-back, and UPDATE behavior under the default setting.

+++ src/metadata_writer_postgres_single.rs
@@ -349,7 +349,7 @@ async fn detect_replace_conflict(
     table_id: i64,
     base_snapshot: i64,
 ) -> Result<()> {
-    let conflict: Option<i64> = sqlx::query(
+    let conflict: Option<i32> = sqlx::query(
         "SELECT 1 FROM ducklake_data_file
          WHERE table_id = $1 AND (begin_snapshot > $2 OR end_snapshot > $2)
          LIMIT 1",
